requirements.txt
requests
python-dateutil
rapidfuzz